# Remove buffer copy when writing base64 JSON value